requirements.txt
asgiref==3.6.0
Django==4.2.5
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
redis==5.2.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.4.0
